requirements.txt
Django==1.7
PyMySQL==0.6.3
South==1.0.1
defusedxml==0.4.1
django-allauth==0.18.0
django-braces==1.4.0
django-configurations==0.8
django-feedreader==1.0.4
django-paypal==0.1.5
django-sslify==0.2.4
djangorestframework==3.0.0
factory-boy==2.3.1
feedparser==5.1.3
mock==1.0.1
oauthlib==0.7.2
python3-openid==3.0.5
pytz==2014.10
requests==2.5.0
requests-oauthlib==0.4.2
six==1.8.0
